feat: update Product interface to use string id and refactor product API slice imports

+++ src/components/TableActions.tsx
@@ -1,28 +1,43 @@
 import { HStack, IconButton } from "@chakra-ui/react";
 import { FaEye, FaEdit, FaTrash } from "react-icons/fa";
+import { Link } from "react-router-dom";
+import AlertDialog from "./ui/AlertDialog";
+import { useDeleteProductMutation } from "../app/Slices/products";
 
 interface TableActionsProps {
-	id: number;
-	onPreview: (id: number) => void;
-	onEdit: (id: number) => void;
-	onDelete: (id: number) => void;
+	id: string;
 }
 
-const TableActions: React.FC<TableActionsProps> = ({
-	id,
-	onPreview,
-	onEdit,
-	onDelete,
-}) => {
+const TableActions: React.FC<TableActionsProps> = ({ id }) => {
+	const [
+		deleteProduct,
+		{
+			isLoading: isDeleteLoading,
+			isError: isDeleteError,
+			isSuccess: isDeleteSuccess,
+		},
+	] = useDeleteProductMutation();
+
+	function onEdit(id: string): void {
+		throw new Error("Function not implemented.");
+	}
+
+	const onDelete = (id: string) => {
+		deleteProduct(id);
+	};
 	return (
 		<HStack spaceX={2}>
-			<IconButton
-				aria-label='Preview'
-				onClick={() => onPreview(id)}
-				variant='outline'
-				colorScheme='blue'>
-				<FaEye />
-			</IconButton>
+			<Link
+				to={`/products/${id}`}
+				target='_blank'
+				rel='noopener noreferrer'>
+				<IconButton
+					aria-label='Preview'
+					variant='outline'
+					colorScheme='blue'>
+					<FaEye />
+				</IconButton>
+			</Link>
 			<IconButton
 				aria-label='Edit'
 				onClick={() => onEdit(id)}
@@ -30,13 +45,22 @@ const TableActions: React.FC<TableActionsProps> = ({
 				colorScheme='yellow'>
 				<FaEdit />
 			</IconButton>
-			<IconButton
-				aria-label='Delete'
-				onClick={() => onDelete(id)}
-				variant='outline'
-				colorScheme='red'>
-				<FaTrash />
-			</IconButton>
+			<AlertDialog
+				title='Are you sure?'
+				description='This action cannot be undone. This will permanently delete your account and remove your data from our systems.'
+				confirmLabel='Delete'
+				cancelLabel='Cancel'
+				onConfirm={() => onDelete(id)}
+				isLoading={isDeleteLoading}
+				isError={isDeleteError}
+				isSuccess={isDeleteSuccess}>
+				<IconButton
+					aria-label='Delete'
+					variant='outline'
+					colorScheme='red'>
+					<FaTrash />
+				</IconButton>
+			</AlertDialog>
 		</HStack>
 	);
 };
